fix: Critical bugs and optimizations from plan.md
- FPS preset logic: fixed search (> instead of >=), fps=25.5 now correctly goes to 30
- Worker count: proper rounding with div_ceil(4) instead of truncating division
- Jog methods: removed duplication, created shared start_jog(direction)
- Frame counting: O(1) AtomicUsize counter instead of O(n) iteration
- System caching: reuse System instance in set_memory_fraction
- Tests: fixed VideoCodec::MPEG4 reference, 14/14 tests passing
- Clippy: auto-fixed 54 warnings

+++ src/video.rs
@@ -90,7 +90,7 @@ pub fn decode_frame(
 
     // Enable multi-threaded frame decoding (2-4x speedup)
     unsafe {
-        (*decoder_ctx.as_mut_ptr()).thread_type = ffmpeg::ffi::FF_THREAD_FRAME as i32;
+        (*decoder_ctx.as_mut_ptr()).thread_type = ffmpeg::ffi::FF_THREAD_FRAME;
         (*decoder_ctx.as_mut_ptr()).thread_count = 0; // Auto-detect CPU cores
     }
